seeders

+++ database/migrations/2025_04_20_132850_services.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->integer('pbs_service_for');
             $table->string('pbs_name');
             $table->text('pbs_description')->nullable();
-            $table->decimal('pbs_duration', 2, 2)->default(0.00);
+            $table->integer('pbs_duration')->default(0);
             $table->string('pbs_image')->nullable();
             $table->float('pbs_price', 6, 2)->default(0.00);
             $table->string('pbs_employees')->nullable();
